requirements.txt
streamlit
pandas
google-generativeai
scikit-learn
joblib
python-docx
PyPDF2
nltk